Refactor backend

+++ server/src/main/java/ai/equity/salt/openai/controller/dto/CompanyOverview.java
@@ -4,10 +4,9 @@ import java.util.List;
 
 public record CompanyOverview(
         List<JobDataSet> topFiveHighestPayingPositions,
-        long totalNumberOfEmployees,
-
-        double averageAge,
-        double averageTenure,
-        double averageSalary
+        int totalNumberOfEmployees,
+        int averageAge,
+        int averageTenure,
+        int averageSalary
 ) {
 }
